requirements.txt
opencv-python
pytesseract
tabula-py
paho-mqtt
QT
PySide6
pyspellchecker


